mm: add zrblock allocator
zblock is a special purpose allocator for storing compressed pages. It
is implemented in Rust language with the focus on safety and
performance. It stores integer number of same size compressed objects
per its block. These blocks consist of several virtually contiguous
physical pages (1 - 8 pages per block).

With zblock, it is possible to densely arrange objects of various sizes
resulting in very high peak compression density, not dependent on the
compaction execution.

Signed-off-by: Vitaly Wool <vitaly.wool@konsulko.se>
Signed-off-by: Alice Ryhl <aliceryhl@google.com>