zblock_rust: use ForeignOwnable in Zpool
Instead of passing raw void pointers around in Rust code, we use the
ForeignOwnable trait which can be used to represent any pointer type
that can be converted to/from a void pointer.

To work with the void pointer, we have three methods:
* into_foreign converts Rust type into void pointer
* from_foreign converts void pointer into Rust type with ownership
* borrow converts void pointer into Rust type without ownership

We use borrow in the cases where the method needs to access the Rust
type stored in the void pointer without taking ownership away from C
code, i.e. when the void pointer remains valid after the call. When
destroying the pool, we instead use from_foreign.

In zblock we specify that the pool is stored inside the pointer type
KBox, and the methods create and destroy get to use the KBox<ZblockPool>
type in the method signatures as they have ownership of the box. On the
other hand, the other methods just get a &ZblockPool instead because a
reference is the borrowed version of KBox<_>. (This is specified in the
`impl ForeignOwnable for Box` block in rust/kernel/alloc/kbox.rs).

Note that we are using borrow() instead of borrow_mut() here. This means
that I had to add `c_pool_as_mutable` as a workaround for now. Using
mutable references is incorrect because the real meaning of mutable in
Rust is "exclusive". Since Zpool::malloc() can be called in parallel,
access to the ZblockPool is *not* exclusive so using &mut ZblockPool is
wrong. (Some people call &T for a shared reference and &mut T for an
exclusive reference instead of using the immutable/mutable wording for
this reason.)

Signed-off-by: Alice Ryhl <aliceryhl@google.com>