zblock_rust: use Rust spinlock
Since Rust enforces that spinlocks are not moved in memory after
creation (they must be "pinned"), we need to change the constructor of
BlockList to construct it in-place so that lockdep is initialized
properly.

Since the KVec type is intended for arrays that are sometimes
reallocated, it doesn't fit the use-case where the contents are pinned
to one memory location. Instead, we use the KBox type for that - I had
to add a new constructor to KBox for our use-case to make this work.

There are some compilation failures related to mutability. Moving the
data inside the spinlock should fix them.

Signed-off-by: Alice Ryhl <aliceryhl@google.com>