zblock_rust: move fields inside spinlock
Moving the fields inside the spinlock gets rid of the mutability errors
again and lets us use Rust's checking that we take a lock before
accessing a field protected by the lock.

Currently the red/black tree implementation only provides cursors for
exclusive access to the rbtree. To get rid of the last use of
c_pool_as_mutable, we will have to add support to the rbtree cursor for
shared access.

Rust now enforces that you must take the spinlock to read the
`block_count`, since we're now saying that the field is protected by the
spinlock. We can fix this by making it an atomic and moving it outside
of the spinlock. If we do this while using relaxed atomics to access it,
then it should compile down to the same code while allowing you to read
it without holding the lock.

Signed-off-by: Alice Ryhl <aliceryhl@google.com>