mm/slub: allow to set node and align in k[v]realloc
Reimplement k[v]realloc() to be able to set node and alignment
should a user need to do so. In order to do that, the following
rules are honored:
* kmalloc/kzalloc/kmalloc_node/kzalloc_node remain unchanged
* kvmalloc_node/kvzalloc_node get an extra parameter (alignment)
* kvcalloc remains unchanged
* kvcalloc_node gets an extra parameter (alignment)
* kvrealloc remains unchanged
* kvrealloc_node is implemented as a new function taking align and
NUMA id as extra parameters compared to kvrealloc.

The users of functions that changed their interfaces are updated
accordingly.

With that change we also provide the ability for the Rust part of
the kernel to set node and aligmnent in its K[v]xxx
[re]allocations.

Signed-off-by: Vitaly Wool <vitaly.wool@konsulko.se>