mm: kconfig: select allocation backend at compile time
With the removal of the zpool layer the opportunity to select an allocation
backend for zswap has been completely lost. While the flexibility zpool
used to provide may have seemed unjustified, the option to select an
alternative backend should still exist, especially with the recent
advancements in the vmap/vmalloc field that allow for faster and simple
allocation backends. This patch adds very simple means for backend compile
time selection.

Signed-off-by: Vitaly Wool <vitaly.wool@konsulko.se>